+++ src/project/infrastructure/postgres/models.py
@@ -1,0 +1,11 @@
+from sqlalchemy.orm import Mapped, mapped_column
+
+from project.infrastructure.postgres.database import Base
+
+
+class Product(Base):
+    __tablename__ = "product"
+
+    id: Mapped[int] = mapped_column(primary_key=True)
+    name: Mapped[str] = mapped_column(nullable=False)
+    cost: Mapped[int] = mapped_column(nullable=False)