+++ Changelog.md
@@ -1,4 +1,9 @@
 ### История изменения проекта по semver (https://semver.org/lang/ru/)
 
+# 0.2.0 (29.10.2024)
+- Добавлены model: product
+- Добавлены CRUD-запросы к product в routes
+- product repo
+
 # 0.1.0 (29.10.2024)
 - Добавлена в schemas таблица product